// src/generated/internal/prismaNamespace.ts
export const VoterRollScalarFieldEnum = {
  id: 'id',
  idType: 'idType',
  idNumber: 'idNumber',
  studentId: 'studentId',
  fullName: 'fullName',
  email: 'email',
  imageUrl: 'imageUrl',
  phone: 'phone',
  campusId: 'campusId',
  facultyId: 'facultyId',
  departmentId: 'departmentId',
  programmeId: 'programmeId',
  level: 'level',
  dateOfBirth: 'dateOfBirth',
  isRegistered: 'isRegistered',
  createdAt: 'createdAt',
  updatedAt: 'updatedAt'
} as const

export type VoterRollScalarFieldEnum = (typeof VoterRollScalarFieldEnum)[keyof typeof VoterRollScalarFieldEnum]